ai_service/requirements.txt
azure-functions
google-genai